margin, font_size changed to float
used to be int

+++ backend/labels/__init__.py
@@ -37,9 +37,9 @@ class SheetSpecifications(pydantic.BaseModel):
     label_width: float
     label_height: float
     corner_radius: int
-    margin: int
+    margin: float
     spacing: int
-    font_size: int
+    font_size: float
     border: bool
     left_margin: float
     right_margin: float
